requirements.txt
streamlit
azure-cognitiveservices-speech
python-dotenv
pandas
numpy
gTTS
psycopg2-binary